Avoid using qapp with autouse=True
